# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Verisafe service API",
        "title": "Verisafe API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Open Crafts Interactive",
            "url": "https://opencrafts.io/about",
            "email": "developers@opencrafts.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "host": "qaverisafe.opencrafts.io",
    "basePath": "/",
    "paths": {
...
        "/permissions": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    },
                    {
                        "ApiKey": []
                    }
                ],
                "description": "Returns a paginated list of every permission defined in the\nsystem. Paginated with limit/offset; the response is a bare\narray, not an envelope.\nRequires the read:permission:any permission.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "permissions"
                ],
                "summary": "List permissions",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page size (default 10, max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page offset (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/repository.Permission"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch permissions",
                        "schema": {
                            "$ref": "#/definitions/core.APIError"
                        }
                    }
                }
            }
        },
        "/permissions/assign/{perm_id}/{role_id}": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    },
                    {
                        "ApiKey": []
                    }
                ],
                "description": "Attaches a permission to a role, granting it to every account\nthat holds the role. Not idempotent: role_permissions is keyed\non (role_id, permission_id), so re-granting a permission the\nrole already holds fails the primary key and returns 500.\nRequires the assign:permission:role permission.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "permissions"
                ],
                "summary": "Grant a permission to a role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Permission ID",
                        "name": "perm_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Role ID",
                        "name": "role_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Confirmation message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid perm_id or role_id",
                        "schema": {
                            "$ref": "#/definitions/core.APIError"
                        }
                    },
                    "500": {
                        "description": "Failed to assign permission",
                        "schema": {
                            "$ref": "#/definitions/core.APIError"
                        }
                    }
                }
            }
        },
        "/permissions/create": {
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    },
                    {
                        "ApiKey": []
                    }
                ],
                "description": "Newly created permissions are automatically granted to the \"system\" and \"Administrator\" roles by database triggers — see docs/RBAC.md.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "permissions"
                ],
                "summary": "Create a permission",
                "parameters": [
                    {
                        "description": "Permission to create",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/repository.CreatePermissionParams"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/repository.Permission"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/core.APIError"
                        }
                    },
                    "500": {
                        "description": "Failed to create permission",
                        "schema": {
                            "$ref": "#/definitions/core.APIError"
                        }
                    }
                }
            }
        },
        "/permissions/revoke/{perm_id}/{role_id}": {
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    },
                    {
                        "ApiKey": []
                    }
                ],
                "description": "Detaches a permission from a role, withdrawing it from every\naccount that held it only through this role. Accounts holding\nthe same permission via another role keep it.\nRequires the revoke:permission:role permission.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "permissions"
                ],
                "summary": "Revoke a permission from a role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Permission ID",
                        "name": "perm_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Role ID",
                        "name": "role_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Confirmation message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid perm_id or role_id",
                        "schema": {
                            "$ref": "#/definitions/core.APIError"
                        }
                    },
                    "500": {
                        "description": "Failed to revoke permission",
                        "schema": {
                            "$ref": "#/definitions/core.APIError"
                        }
                    }
                }
            }
        },
        "/permissions/user/{id}": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    },
                    {
                        "ApiKey": []
                    }
                ],
                "description": "Returns the permissions the account effectively holds, resolved\nthrough every role assigned to it. Permissions are never\nattached to an account directly, only via a role.\nRequires the read:permission:user permission.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "permissions"
                ],
                "summary": "List a given user's effective permissions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/repository.UserPermissionsView"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid id",
                        "schema": {
                            "$ref": "#/definitions/core.APIError"
                        }
                    },
                    "500": {
                        "description": "Failed to fetch permissions",
                        "schema": {
                            "$ref": "#/definitions/core.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "core.APIError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "the provided input is invalid or malformed"
                }
            }
        },
...
        "pgtype.InfinityModifier": {
            "type": "integer",
            "format": "int32",
            "enum": [
                1,
                0,
                -1
            ],
            "x-enum-varnames": [
                "Infinity",
                "Finite",
                "NegativeInfinity"
            ]
        },
        "pgtype.Timestamp": {
            "type": "object",
            "properties": {
                "infinityModifier": {
                    "$ref": "#/definitions/pgtype.InfinityModifier"
                },
                "time": {
                    "description": "Time zone will be ignored when encoding to PostgreSQL.",
                    "type": "string"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        },
...
        "repository.CreatePermissionParams": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "repository.Permission": {
            "type": "object",
            "properties": {
                "created_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                }
            }
        },
...
        "repository.UserPermissionsView": {
            "type": "object",
            "properties": {
                "permission": {
                    "type": "string"
                },
                "permission_id": {
                    "type": "string"
                },
                "role_id": {
                    "type": "string"
                },
                "role_name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKey": {
            "type": "apiKey",
            "name": "X-Api-Key",
            "in": "header"
        },
        "BearerToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}