# internal/handlers/permission/handler.go
package permission

import (
	"database/sql"
	"encoding/json"
	"errors"
	"log/slog"
	"net/http"

	"github.com/google/uuid"
	"github.com/opencrafts-io/verisafe/internal/config"
	"github.com/opencrafts-io/verisafe/internal/core"
	"github.com/opencrafts-io/verisafe/internal/middleware"
	"github.com/opencrafts-io/verisafe/internal/repository"
)

type PermissionHandler struct {
	Cacher core.Cacher
	DB     core.IDBProvider
	Cfg    *config.Config
	Logger *slog.Logger
}

// Registers all the necessary routes associated with this handler group
func (ph *PermissionHandler) RegisterHandlers(
	router *http.ServeMux,
) {
	router.Handle("POST /permissions/create",
		middleware.CreateStack(
			middleware.IsAuthenticated(ph.Cfg, ph.DB, ph.Cacher, ph.Logger),
			middleware.HasPermission([]string{"create:permission"}),
		)(http.HandlerFunc(ph.CreatePermission)),
	)

	router.Handle("GET /permissions",
		middleware.CreateStack(
			middleware.IsAuthenticated(ph.Cfg, ph.DB, ph.Cacher, ph.Logger),
			middleware.HasPermission([]string{"read:permission:any"}),
			middleware.PaginationMiddleware(10, 100),
		)(http.HandlerFunc(ph.GetAllPermissions)),
	)

	router.Handle("GET /permissions/{id}",
		middleware.CreateStack(
			middleware.IsAuthenticated(ph.Cfg, ph.DB, ph.Cacher, ph.Logger),
			middleware.HasPermission([]string{"read:permission:any"}),
		)(http.HandlerFunc(ph.GetPermissionByID)),
	)

	router.Handle("GET /permissions/user/{id}",
		middleware.CreateStack(
			middleware.IsAuthenticated(ph.Cfg, ph.DB, ph.Cacher, ph.Logger),
			middleware.HasPermission([]string{"read:permission:user"}),
		)(http.HandlerFunc(ph.GetAllUserPermissions)),
	)

	router.Handle("PATCH /permissions/{id}",
		middleware.CreateStack(
			middleware.IsAuthenticated(ph.Cfg, ph.DB, ph.Cacher, ph.Logger),
			middleware.HasPermission([]string{"update:permission:any"}),
		)(http.HandlerFunc(ph.UpdatePermission)),
	)

	router.Handle("GET /permissions/assign/{perm_id}/{role_id}",
		middleware.CreateStack(
			middleware.IsAuthenticated(ph.Cfg, ph.DB, ph.Cacher, ph.Logger),
			middleware.HasPermission([]string{"assign:permission:role"}),
		)(http.HandlerFunc(ph.AssignRolePermission)),
	)

	router.Handle("DELETE /permissions/revoke/{perm_id}/{role_id}",
		middleware.CreateStack(
			middleware.IsAuthenticated(ph.Cfg, ph.DB, ph.Cacher, ph.Logger),
			middleware.HasPermission([]string{"revoke:permission:role"}),
		)(http.HandlerFunc(ph.RevokeRolePermission)),
	)
}

// CreatePermission godoc
//
// @Summary      Create a permission
// @Description  Newly created permissions are automatically granted to the "system" and "Administrator" roles by database triggers — see docs/RBAC.md.
// @Tags         permissions
// @Accept       json
// @Produce      json
// @Param        request  body      repository.CreatePermissionParams  true  "Permission to create"
// @Success      201  {object}  repository.Permission
// @Failure      400  {object}  core.APIError  "Invalid request body"
// @Failure      500  {object}  core.APIError  "Failed to create permission"
// @Security     BearerToken
// @Security     ApiKey
// @Router       /permissions/create [post]
func (ph *PermissionHandler) CreatePermission(
	w http.ResponseWriter,
	r *http.Request,
) {
	w.Header().Set("Content-Type", "application/json")
	conn, err := middleware.GetDBConnFromContext(r.Context())
	if err != nil {
		ph.Logger.Error(
			"Error while processing request",
			slog.Any("error", err),
		)
		w.WriteHeader(http.StatusInternalServerError)
		json.NewEncoder(w).Encode(map[string]string{
			"error": "We ran into a problem while servicing your request please try again later",
		})
		return
	}

	tx, _ := conn.Begin(r.Context())
	defer tx.Rollback(r.Context())
	repo := repository.New(tx)

	var permData repository.CreatePermissionParams

	if err := json.NewDecoder(r.Body).Decode(&permData); err != nil {
		ph.Logger.Error("Failed to parse request body", slog.Any("error", err))
		w.WriteHeader(http.StatusBadRequest)
		json.NewEncoder(w).Encode(map[string]string{
			"error": "Please check your request body and try again",
		})
		return
	}

	created, err := repo.CreatePermission(r.Context(), permData)
	if err != nil {
		ph.Logger.Error("Failed to create permission",
			slog.Any("error", err),
			slog.Any("permission", permData),
		)
		w.WriteHeader(http.StatusInternalServerError)
		json.NewEncoder(w).Encode(map[string]string{
			"error": "We couldn't create this permission at the moment please try again later",
		})
		return
	}

	if err = tx.Commit(r.Context()); err != nil {
		ph.Logger.Error(
			"Error while committing transaction",
			slog.Any("error", err),
		)
		w.WriteHeader(http.StatusInternalServerError)
		json.NewEncoder(w).Encode(map[string]any{
			"error": "We ran into a problem while servicing your request please try again later",
		})
		return
	}

	w.WriteHeader(http.StatusCreated)
	json.NewEncoder(w).Encode(created)
}

// GetPermissionByID godoc
//
// @Summary      Get a permission by id
// @Description  Returns a single permission by its UUID. Responds 404 when no
// @Description  permission carries that id.
// @Description  Requires the read:permission:any permission.
// @Tags         permissions
// @Produce      json
// @Param        id  path  string  true  "Permission ID"
// @Success      200  {object}  repository.Permission
// @Failure      400  {object}  core.APIError  "Invalid id"
// @Failure      404  {object}  core.APIError  "Permission not found"
// @Failure      500  {object}  core.APIError  "Failed to fetch permission"
// @Security     BearerToken
// @Security     ApiKey
// @Router       /permissions/{id} [get]
func (ph *PermissionHandler) GetPermissionByID(
	w http.ResponseWriter,
	r *http.Request,
) {
	rawID := r.PathValue("id")
	id, err := uuid.Parse(rawID)
	if err != nil {
		ph.Logger.Error("Failed to parse request body", slog.Any("error", err))
		w.WriteHeader(http.StatusBadRequest)
		json.NewEncoder(w).Encode(map[string]string{
			"error": "Please check your request body and try again",
		})
		return
	}

	w.Header().Set("Content-Type", "application/json")
	conn, err := middleware.GetDBConnFromContext(r.Context())
	if err != nil {
		ph.Logger.Error(
			"Error while processing request",
			slog.Any("error", err),
		)
		w.WriteHeader(http.StatusInternalServerError)
		json.NewEncoder(w).Encode(map[string]string{
			"error": "We ran into a problem while servicing your request please try again later",
		})
		return
	}

	tx, _ := conn.Begin(r.Context())
	defer tx.Rollback(r.Context())
	repo := repository.New(tx)

	role, err := repo.GetPermissionByID(r.Context(), id)
	if errors.Is(err, sql.ErrNoRows) {
		w.WriteHeader(http.StatusNotFound)
		json.NewEncoder(w).Encode(map[string]string{
			"error": "The permission you are requesting does not exist",
		})
		return
	}
	if err != nil {
		ph.Logger.Error("Failed to retrieve permission",
			slog.Any("error", err), slog.Any("role", id.String()),
		)
		w.WriteHeader(http.StatusInternalServerError)
		json.NewEncoder(w).Encode(map[string]string{
			"error": "We couldn't complete this request at the moment please try again later",
		})
		return
	}

	w.WriteHeader(http.StatusOK)
	json.NewEncoder(w).Encode(role)
}

// GetAllPermissions godoc
//
// @Summary      List permissions
// @Description  Returns a paginated list of every permission defined in the
// @Description  system. Paginated with limit/offset; the response is a bare
// @Description  array, not an envelope.
// @Description  Requires the read:permission:any permission.
// @Tags         permissions
// @Produce      json
// @Param        limit   query  int  false  "Page size (default 10, max 100)"
// @Param        offset  query  int  false  "Page offset (default 0)"
// @Success      200  {array}   repository.Permission
// @Failure      500  {object}  core.APIError  "Failed to fetch permissions"
// @Security     BearerToken
// @Security     ApiKey
// @Router       /permissions [get]
func (ph *PermissionHandler) GetAllPermissions(
	w http.ResponseWriter,
	r *http.Request,
) {
	pagination := middleware.GetPagination(r.Context())

	w.Header().Set("Content-Type", "application/json")
	conn, err := middleware.GetDBConnFromContext(r.Context())
	if err != nil {
		ph.Logger.Error(
			"Error while processing request",
			slog.Any("error", err),
		)
		w.WriteHeader(http.StatusInternalServerError)
		json.NewEncoder(w).Encode(map[string]string{
			"error": "We ran into a problem while servicing your request please try again later",
		})
		return
	}

	tx, _ := conn.Begin(r.Context())
	defer tx.Rollback(r.Context())
	repo := repository.New(tx)

	roles, err := repo.GetAllPermissions(
		r.Context(),
		repository.GetAllPermissionsParams{
			Limit:  int32(pagination.Limit),
			Offset: int32(pagination.Offset),
		},
	)
	if err != nil {
		ph.Logger.Error(
			"Failed to retrieve permissions",
			slog.Any("error", err),
		)
		w.WriteHeader(http.StatusInternalServerError)
		json.NewEncoder(w).Encode(map[string]string{
			"error": "We couldn't complete this request at the moment please try again later",
		})
		return
	}

	w.WriteHeader(http.StatusOK)
	json.NewEncoder(w).Encode(roles)
}

// GetAllUserPermissions godoc
//
// @Summary      List a given user's effective permissions
// @Description  Returns the permissions the account effectively holds, resolved
// @Description  through every role assigned to it. Permissions are never
// @Description  attached to an account directly, only via a role.
// @Description  Requires the read:permission:user permission.
// @Tags         permissions
// @Produce      json
// @Param        id  path  string  true  "Account ID"
// @Success      200  {array}   repository.UserPermissionsView
// @Failure      400  {object}  core.APIError  "Invalid id"
// @Failure      500  {object}  core.APIError  "Failed to fetch permissions"
// @Security     BearerToken
// @Security     ApiKey
// @Router       /permissions/user/{id} [get]
func (ph *PermissionHandler) GetAllUserPermissions(
	w http.ResponseWriter,
	r *http.Request,
) {
	rawID := r.PathValue("id")
	id, err := uuid.Parse(rawID)
	if err != nil {
		ph.Logger.Error("Failed to parse request body", slog.Any("error", err))
		w.WriteHeader(http.StatusBadRequest)
		json.NewEncoder(w).Encode(map[string]string{
			"error": "Please check your request body and try again",
		})
		return
	}

	w.Header().Set("Content-Type", "application/json")
	conn, err := middleware.GetDBConnFromContext(r.Context())
	if err != nil {
		ph.Logger.Error(
			"Error while processing request",
			slog.Any("error", err),
		)
		w.WriteHeader(http.StatusInternalServerError)
		json.NewEncoder(w).Encode(map[string]string{
			"error": "We ran into a problem while servicing your request please try again later",
		})
		return
	}

	tx, _ := conn.Begin(r.Context())
	defer tx.Rollback(r.Context())
	repo := repository.New(tx)

	roles, err := repo.GetUserPermissions(r.Context(), id)
	if err != nil && !errors.Is(err, sql.ErrNoRows) {
		ph.Logger.Error(
			"Failed to retrieve permissions",
			slog.Any("error", err),
		)
		w.WriteHeader(http.StatusInternalServerError)
		json.NewEncoder(w).Encode(map[string]any{
			"error": "We ran into an issue while retrieving this user's permissions try again later",
		})
		return
	}

	w.WriteHeader(http.StatusOK)
	json.NewEncoder(w).Encode(roles)
}

// UpdatePermission godoc
//
// @Summary      Update a permission
// @Description  Updates a permission in place. Every role holding it, and every
// @Description  account holding those roles, sees the change immediately.
// @Description  Requires the update:permission:any permission.
// @Tags         permissions
// @Accept       json
// @Produce      json
// @Param        request  body      repository.UpdatePermissionParams  true  "Fields to update (id identifies the permission)"
// @Success      200  {object}  repository.Permission
// @Failure      400  {object}  core.APIError  "Invalid request body"
// @Failure      500  {object}  core.APIError  "Failed to update permission"
// @Security     BearerToken
// @Security     ApiKey
// @Router       /permissions/{id} [patch]
func (ph *PermissionHandler) UpdatePermission(
	w http.ResponseWriter,
	r *http.Request,
) {
	w.Header().Set("Content-Type", "application/json")
	conn, err := middleware.GetDBConnFromContext(r.Context())
	if err != nil {
		ph.Logger.Error(
			"Error while processing request",
			slog.Any("error", err),
		)
		w.WriteHeader(http.StatusInternalServerError)
		json.NewEncoder(w).Encode(map[string]string{
			"error": "We ran into a problem while servicing your request please try again later",
		})
		return
	}

	tx, _ := conn.Begin(r.Context())
	defer tx.Rollback(r.Context())
	repo := repository.New(tx)

	var permData repository.UpdatePermissionParams

	if err := json.NewDecoder(r.Body).Decode(&permData); err != nil {
		ph.Logger.Error("Failed to parse request body", slog.Any("error", err))
		w.WriteHeader(http.StatusBadRequest)
		json.NewEncoder(w).Encode(map[string]string{
			"error": "Please check your request body and try again",
		})
		return
	}

	created, err := repo.UpdatePermission(r.Context(), permData)
	if err != nil {
		ph.Logger.Error("Failed to update permission",
			slog.Any("error", err),
			slog.Any("permission", permData),
		)
		w.WriteHeader(http.StatusInternalServerError)
		json.NewEncoder(w).Encode(map[string]string{
			"error": "We couldn't complete this request at the moment please try again",
		})
		return
	}

	if err = tx.Commit(r.Context()); err != nil {
		ph.Logger.Error(
			"Error while committing transaction",
			slog.Any("error", err),
		)
		w.WriteHeader(http.StatusInternalServerError)
		json.NewEncoder(w).Encode(map[string]any{
			"error": "We ran into a problem while servicing your request please try again later",
		})
		return
	}

	w.WriteHeader(http.StatusOK)
	json.NewEncoder(w).Encode(created)
}

// AssignRolePermission godoc
//
// @Summary      Grant a permission to a role
// @Description  Attaches a permission to a role, granting it to every account
// @Description  that holds the role. Not idempotent: role_permissions is keyed
// @Description  on (role_id, permission_id), so re-granting a permission the
// @Description  role already holds fails the primary key and returns 500.
// @Description  Requires the assign:permission:role permission.
// @Tags         permissions
// @Produce      json
// @Param        perm_id  path  string  true  "Permission ID"
// @Param        role_id  path  string  true  "Role ID"
// @Success      200  {object}  map[string]any  "Confirmation message"
// @Failure      400  {object}  core.APIError  "Invalid perm_id or role_id"
// @Failure      500  {object}  core.APIError  "Failed to assign permission"
// @Security     BearerToken
// @Security     ApiKey
// @Router       /permissions/assign/{perm_id}/{role_id} [get]
func (ph *PermissionHandler) AssignRolePermission(
	w http.ResponseWriter,
	r *http.Request,
) {
	rawRoleID := r.PathValue("role_id")
	roleID, err := uuid.Parse(rawRoleID)
	if err != nil {
		ph.Logger.Error("Failed to parse request body", slog.Any("error", err))
		w.WriteHeader(http.StatusBadRequest)
		json.NewEncoder(w).Encode(map[string]string{
			"error": "Please check your request body and try again",
		})
		return
	}

	rawPermID := r.PathValue("perm_id")
	permID, err := uuid.Parse(rawPermID)
	if err != nil {
		ph.Logger.Error("Failed to parse request body", slog.Any("error", err))
		w.WriteHeader(http.StatusBadRequest)
		json.NewEncoder(w).Encode(map[string]string{
			"error": "Please check your request body and try again",
		})
		return
	}

	w.Header().Set("Content-Type", "application/json")
	conn, err := middleware.GetDBConnFromContext(r.Context())
	if err != nil {
		ph.Logger.Error(
			"Error while processing request",
			slog.Any("error", err),
		)
		w.WriteHeader(http.StatusInternalServerError)
		json.NewEncoder(w).Encode(map[string]string{
			"error": "We ran into a problem while servicing your request please try again later",
		})
		return
	}

	tx, _ := conn.Begin(r.Context())
	defer tx.Rollback(r.Context())
	repo := repository.New(tx)

	_, err = repo.AssignRolePermission(
		r.Context(),
		repository.AssignRolePermissionParams{
			PermissionID: permID,
			RoleID:       roleID,
		},
	)
	if err != nil {
		ph.Logger.Error("Failed to assign permission to role",
			slog.Any("error", err),
			slog.Any("role", roleID.String()),
			slog.Any("permission", permID.String()),
		)
		w.WriteHeader(http.StatusInternalServerError)
		json.NewEncoder(w).Encode(map[string]string{
			"error": "We couldn't complete this request at the moment please try again later",
		})
		return
	}

	if err = tx.Commit(r.Context()); err != nil {
		ph.Logger.Error(
			"Error while committing transaction",
			slog.Any("error", err),
		)
		w.WriteHeader(http.StatusInternalServerError)
		json.NewEncoder(w).Encode(map[string]any{
			"error": "We ran into a problem while servicing your request please try again later",
		})
		return
	}

	w.WriteHeader(http.StatusOK)
	json.NewEncoder(w).
		Encode(map[string]any{"message": "Permission successfully assigned"})
}

// RevokeRolePermission godoc
//
// @Summary      Revoke a permission from a role
// @Description  Detaches a permission from a role, withdrawing it from every
// @Description  account that held it only through this role. Accounts holding
// @Description  the same permission via another role keep it.
// @Description  Requires the revoke:permission:role permission.
// @Tags         permissions
// @Produce      json
// @Param        perm_id  path  string  true  "Permission ID"
// @Param        role_id  path  string  true  "Role ID"
// @Success      200  {object}  map[string]any  "Confirmation message"
// @Failure      400  {object}  core.APIError  "Invalid perm_id or role_id"
// @Failure      500  {object}  core.APIError  "Failed to revoke permission"
// @Security     BearerToken
// @Security     ApiKey
// @Router       /permissions/revoke/{perm_id}/{role_id} [delete]
func (ph *PermissionHandler) RevokeRolePermission(
	w http.ResponseWriter,
	r *http.Request,
) {
	rawRoleID := r.PathValue("role_id")
	roleID, err := uuid.Parse(rawRoleID)
	if err != nil {
		ph.Logger.Error("Failed to parse request body", slog.Any("error", err))
		w.WriteHeader(http.StatusBadRequest)
		json.NewEncoder(w).Encode(map[string]string{
			"error": "Please check your request body and try again",
		})
		return
	}

	rawPermID := r.PathValue("perm_id")
	permID, err := uuid.Parse(rawPermID)
	if err != nil {
		ph.Logger.Error("Failed to parse request body", slog.Any("error", err))
		w.WriteHeader(http.StatusBadRequest)
		json.NewEncoder(w).Encode(map[string]string{
			"error": "Please check your request body and try again",
		})
		return
	}

	w.Header().Set("Content-Type", "application/json")
	conn, err := middleware.GetDBConnFromContext(r.Context())
	if err != nil {
		ph.Logger.Error(
			"Error while processing request",
			slog.Any("error", err),
		)
		w.WriteHeader(http.StatusInternalServerError)
		json.NewEncoder(w).Encode(map[string]string{
			"error": "We ran into a problem while servicing your request please try again later",
		})
		return
	}

	tx, _ := conn.Begin(r.Context())
	defer tx.Rollback(r.Context())
	repo := repository.New(tx)

	err = repo.RevokeRolePermission(
		r.Context(),
		repository.RevokeRolePermissionParams{
			PermissionID: permID,
			RoleID:       roleID,
		},
	)
	if err != nil {
		ph.Logger.Error("Failed to revoke permission from role",
			slog.Any("error", err),
			slog.Any("role", roleID.String()),
			slog.Any("permission", permID.String()),
		)
		w.WriteHeader(http.StatusInternalServerError)
		json.NewEncoder(w).Encode(map[string]string{
			"error": "We couldn't complete this request at the moment please try again later",
		})
		return
	}

	if err = tx.Commit(r.Context()); err != nil {
		ph.Logger.Error(
			"Error while committing transaction",
			slog.Any("error", err),
		)
		w.WriteHeader(http.StatusInternalServerError)
		json.NewEncoder(w).Encode(map[string]any{
			"error": "We ran into a problem while servicing your request please try again later",
		})
		return
	}

	w.WriteHeader(http.StatusOK)
	json.NewEncoder(w).
		Encode(map[string]any{"message": "Permission successfully revoked from role"})
}